move nonempty type to publicly accessible location
