+++ fluree-db-api/src/revert.rs
@@ -26,7 +26,7 @@ use fluree_db_core::commit::{TxnMetaEntry, TxnMetaValue};
 use fluree_db_core::ledger_id::format_ledger_id;
 use fluree_db_core::{
     collect_dag_cids, load_commit_by_id, load_commit_envelope_by_id, trace_commits_by_id,
-    BranchedContentStore, CommitId, ConflictKey, ContentStore,
+    BranchedContentStore, CommitId, ConflictKey, ContentStore, NonEmpty,
 };
 use fluree_db_ledger::{LedgerState, StagedLedger};
 use fluree_db_nameservice::NsRecordSnapshot;
@@ -417,51 +417,6 @@ impl crate::Fluree {
 // ---------------------------------------------------------------------------
 // Internal types
 // ---------------------------------------------------------------------------
-
-/// Sequence with a type-level guarantee of at least one element. The
-/// invariant is structural — `head` is always present — so downstream code
-/// can rely on `first`, `iter`, etc. without empty-checks. Constructed only
-/// at validation boundaries.
-#[derive(Clone, Debug)]
-pub(crate) struct NonEmpty<T> {
-    head: T,
-    tail: Vec<T>,
-}
-
-impl<T> NonEmpty<T> {
-    pub(crate) fn try_from_vec(v: Vec<T>) -> Option<Self> {
-        let mut iter = v.into_iter();
-        let head = iter.next()?;
-        Some(Self {
-            head,
-            tail: iter.collect(),
-        })
-    }
-
-    pub(crate) fn iter(&self) -> impl DoubleEndedIterator<Item = &T> {
-        std::iter::once(&self.head).chain(self.tail.iter())
-    }
-
-    pub(crate) fn into_vec(self) -> Vec<T> {
-        let mut v = Vec::with_capacity(1 + self.tail.len());
-        v.push(self.head);
-        v.extend(self.tail);
-        v
-    }
-
-    pub(crate) fn len(&self) -> usize {
-        1 + self.tail.len()
-    }
-}
-
-impl<T> From<T> for NonEmpty<T> {
-    fn from(t: T) -> Self {
-        Self {
-            head: t,
-            tail: Vec::new(),
-        }
-    }
-}
 
 /// Caller-supplied source of the commit list, with [`CommitRef`]s still
 /// unresolved.
